requirements.txt
bleach==6.2.0
dataclass-wizard==0.27.0
fastapi==0.115.5
langchain-core==0.3.18
langchain-nvidia-ai-endpoints==0.3.7
langchain-openai==0.2.8
langchain==0.3.7
langchain-community==0.3.7
pymilvus==2.5.4
python-multipart==0.0.18
unstructured[all-docs]==0.16.11
uvicorn[standard]==0.32.0
pydantic==2.9.2
PyYAML==6.0.2
langchain-milvus==0.1.8
minio==7.2.15
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-prometheus==0.50b0
opentelemetry-instrumentation-milvus==0.36.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-processor-baggage==0.50b0
gradio>=4.0.0
